optimize struct
