+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/table/functions/TemporalTableFunction.scala
@@ -34,10 +34,10 @@ import org.apache.flink.types.Row
   * into other operators (like Temporal Table Join).
   */
 class TemporalTableFunction private(
-                                     @transient private val underlyingHistoryTable: Table,
-                                     private val timeAttribute: PlannerExpression,
-                                     private val primaryKey: String,
-                                     private val resultType: RowTypeInfo)
+    @transient private val underlyingHistoryTable: Table,
+    private val timeAttribute: PlannerExpression,
+    private val primaryKey: String,
+    private val resultType: RowTypeInfo)
   extends TableFunction[Row] {
 
   def eval(row: Timestamp): Unit = {
@@ -66,9 +66,9 @@ class TemporalTableFunction private(
 
 object TemporalTableFunction {
   private[flink] def create(
-                             table: Table,
-                             timeAttribute: PlannerExpression,
-                             primaryKey: String): TemporalTableFunction = {
+      table: Table,
+      timeAttribute: PlannerExpression,
+      primaryKey: String): TemporalTableFunction = {
     new TemporalTableFunction(
       table,
       timeAttribute,
