+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/table/runtime/aggregate/AggregateUtil.scala
@@ -1639,9 +1639,9 @@ object AggregateUtil {
 
 
   private def getTimeFieldPosition(
-                                    timeField: PlannerExpression,
-                                    inputType: RelDataType,
-                                    isParserCaseSensitive: Boolean): Int = {
+      timeField: PlannerExpression,
+      inputType: RelDataType,
+      isParserCaseSensitive: Boolean): Int = {
 
     timeField match {
       case PlannerResolvedFieldReference(name, _) =>
